requirements.txt
google-generativeai>=0.3.0
transformers>=4.30.0
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.21.0
accelerate>=0.21.0
moviepy>=1.0.3
tensorflow>=2.12.0
tensorboard>=2.12.0
opencv-python>=4.7.0
controlnet-aux>=0.0.6
ftfy>=6.1.1
scipy>=1.10.1
tqdm>=4.65.0
pydrive>=1.3.1
gradio>=3.36.1
numpy>=1.24.3
Pillow>=9.5.0
matplotlib>=3.7.2
python-dotenv>=1.0.0
omegaconf>=2.3.0
einops>=0.6.1
xformers>=0.0.20
pyngrok>=6.0.0
bark>=0.0.1
TTS>=0.17.6
ffmpeg-python>=0.2.0
rife>=0.0.1
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.97.0
PyYAML>=6.0.1
sentencepiece>=0.1.99
animatediff>=1.0.0
